requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiosqlite>=0.19.0
pillow>=10.2.0
humanfriendly>=10.0
pytz>=2024.1
psutil>=7.0.0 